+++ company/src/main/java/com/hubEleven/company/presentation/controller/CompanyController.java
@@ -9,7 +9,7 @@ import com.hubEleven.company.application.command.CreateCompanyCommand;
 import com.hubEleven.company.application.command.SearchCompanyCommand;
 import com.hubEleven.company.application.command.UpdateCompanyCommand;
 import com.hubEleven.company.application.dto.response.CompanyResult;
-import com.hubEleven.company.application.service.CompanyAppService;
+import com.hubEleven.company.application.service.CompanyService;
 import com.hubEleven.company.domain.vo.CompanyStatus;
 import com.hubEleven.company.domain.vo.CompanyType;
 import com.hubEleven.company.presentation.dto.request.ChangeCompanyStatusRequest;
@@ -30,7 +30,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/v1/companies")
 public class CompanyController {
 
-	private final CompanyAppService companyAppService;
+	private final CompanyService companyService;
 
 	@Operation(summary = "업체 생성 API", description = "새로운 업체를 생성한다.")
 	@PostMapping
@@ -40,7 +40,7 @@ public class CompanyController {
 			@RequestHeader("X-User-Role") String userRole) {
 
 		CompanyResult result =
-				companyAppService.createCompany(
+				companyService.createCompany(
 						new CreateCompanyCommand(
 								req.hubId(), req.name(), req.type(), req.slackId(), req.address()),
 						userId,
@@ -57,7 +57,7 @@ public class CompanyController {
 			@RequestHeader("X-User-Role") String userRole) {
 
 		CompanyResult result =
-				companyAppService.updateCompany(
+				companyService.updateCompany(
 						new UpdateCompanyCommand(
 								companyId, req.name(), req.type(), req.slackId(), req.address()),
 						userId,
@@ -68,7 +68,7 @@ public class CompanyController {
 	@Operation(summary = "업체 단건 조회 API", description = "업체 ID로 업체를 조회한다.")
 	@GetMapping("/{companyId}")
 	public ResponseEntity<ApiResponse<CompanyResponse>> getCompany(@PathVariable UUID companyId) {
-		CompanyResult dto = companyAppService.getCompany(companyId);
+		CompanyResult dto = companyService.getCompany(companyId);
 		return ApiResponseEntity.success(CompanyResponse.from(dto));
 	}
 
@@ -76,7 +76,7 @@ public class CompanyController {
 	@GetMapping
 	public ResponseEntity<ApiResponse<CommonPageResponse<CompanyResponse>>> getCompanyList(
 			@Valid CommonPageRequest pageReq) {
-		var page = companyAppService.findCompanyList(pageReq);
+		var page = companyService.findCompanyList(pageReq);
 		var mapped =
 				new CommonPageResponse<>(
 						page.content().stream().map(CompanyResponse::from).toList(),
@@ -99,7 +99,7 @@ public class CompanyController {
 			@RequestParam(required = false) CompanyStatus status) {
 
 		var page =
-				companyAppService.searchCompany(
+				companyService.searchCompany(
 						new SearchCompanyCommand(hubId, name, type, status), pageReq);
 
 		var mapped =
@@ -123,7 +123,7 @@ public class CompanyController {
 			@RequestHeader("X-User-Role") String userRole) {
 
 		CompanyResult result =
-				companyAppService.changeStatus(
+				companyService.changeStatus(
 						new ChangeCompanyStatusCommand(companyId, req.status()), userId, userRole);
 		return ApiResponseEntity.success(CompanyResponse.from(result));
 	}
@@ -135,7 +135,7 @@ public class CompanyController {
 			@RequestHeader("X-User-Id") Long userId,
 			@RequestHeader("X-User-Role") String userRole) {
 
-		companyAppService.deleteCompany(companyId, userId, userRole);
+		companyService.deleteCompany(companyId, userId, userRole);
 		return ApiResponseEntity.success(null);
 	}
 }
